+++ main-service/src/main/java/ru/practicum/ewm/dto/compilation/GetManyCompilationDto.java
@@ -7,6 +7,6 @@ import lombok.Value;
 @Builder
 public class GetManyCompilationDto {
     Boolean pinned;
-    Long from;
-    Long size;
+    Integer from;
+    Integer size;
 }
